+++ src/commands/mod.rs
@@ -1,12 +1,13 @@
+pub mod calculate;
+pub mod config;
 pub mod database;
 pub mod devices;
 pub mod info;
 pub mod summary;
-pub mod calculate;
 
+pub use calculate::CalculateArgs;
+pub use config::ConfigArgs;
 pub use database::DatabaseArgs;
 pub use devices::DevicesArgs;
 pub use info::InfoArgs;
 pub use summary::SummaryArgs;
-pub use calculate::CalculateArgs;
-